+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Added a way to get stats about bundles (`FlashbotsMiddleware::get_bundle_stats`)
+- Added a getter for the bundle hash of a pending bundle
 
 ## [0.3.0]